+++ Cavern.Filters/DebugCrossover.cs
@@ -5,7 +5,7 @@
         /// <param name="sampleRate">Audio sample rate</param>
         /// <param name="frequency">Crossover frequency</param>
         /// <param name="order">Number of filters per pass, 2 is recommended for mixing notch prevention</param>
-        public DebugCrossover(int sampleRate, float frequency, int order = 2) : base(sampleRate, frequency, order) { }
+        public DebugCrossover(int sampleRate, double frequency, int order = 2) : base(sampleRate, frequency, order) { }
 
         /// <summary>Apply crossover on an array of samples. One filter should be applied to only one continuous stream of samples.</summary>
         public override void Process(float[] samples) {
